Implemented a simple blacklist

+++ scoreboard.php
@@ -1,8 +1,5 @@
 <?php
 
-header("Content-Type: application/json; charset: utf-8");
-
-
 // Database configuration
 $host = "localhost";
 $user = "score";
@@ -23,6 +20,7 @@ if (mysqli_connect_errno($dblink)) {
 
 
 switch ($_SERVER["REQUEST_METHOD"]) {
+	
 	case "GET":
 		
 		// Query the top 4 scores from the database
@@ -53,30 +51,63 @@ switch ($_SERVER["REQUEST_METHOD"]) {
 			);
 		}
 		
+		header("Content-Type: application/json; charset: utf-8");
+		
 		// Encode the top5 list in JSON
 		echo json_encode($top5, JSON_UNESCAPED_UNICODE);
 		
 		break;
 	
+	
 	case "POST":
+		
+		// Get the user IP
+		$ip = ip2long($_SERVER["REMOTE_ADDR"]);
+		
+		// Check if it is blacklisted
+		$listed = mysqli_query($dblink, "SELECT * FROM `blacklist` WHERE IP='".$ip."' LIMIT 1");
+		
+		if (mysqli_fetch_row($listed)) {
+			
+			http_response_code(403);
+			die("Stahp!");
+		}
 		
 		// Sanitize the inputs
 		$name = mysqli_real_escape_string($dblink, $_POST["name"]);
 		$score = mysqli_real_escape_string($dblink, $_POST["score"]);
 		
 		
-		// Get the best score by the player from the database
-		$oldscore = mysqli_fetch_assoc(mysqli_query($dblink, "SELECT score FROM `".$table."` WHERE name='".$name."' ORDER by score DESC LIMIT 1"))["score"];
+		// Check if the inputs are invalid
+		if ($name == '' | $name == 'Ossi Portaankorva') {
+
+			http_response_code(400);
+			die();
+		}
 		
+		// Blacklist the IP if the score is too high
+		if ($score > 625) {
+			
+			mysqli_query($dblink, "INSERT INTO `blacklist` (`IP`) VALUES ('".$ip."') ");
+			http_response_code(400);
+			die("No Way!");
+		}
 		
-		// Insert the new score to the database only if it is better than the old one
-		if ($score > $oldscore) {
+		if ($score>=5) {
 			
-			// Insert the new score to the database
-			mysqli_query($dblink, "INSERT INTO `".$table."` (`name`, `score`) VALUES ('".$name."','".$score."')");
+			// Get the best score by the player from the database
+			$oldscore = mysqli_fetch_assoc(mysqli_query($dblink, "SELECT score FROM `".$table."` WHERE name='".$name."' ORDER by score DESC LIMIT 1"))["score"];
 			
-			// Delete the player's old score
-			mysqli_query($dblink, "DELETE FROM `".$table."` WHERE name='".$name."' AND score<'".$score."'");
+			
+			// Insert the new score to the database only if it is better than the old one
+			if ($score > $oldscore) {
+				
+				// Insert the new score to the database
+				mysqli_query($dblink, "INSERT INTO `".$table."` (`name`, `score`) VALUES ('".$name."','".$score."')");
+				
+				// Delete the player's old score
+				mysqli_query($dblink, "DELETE FROM `".$table."` WHERE name='".$name."' AND score<'".$score."'");
+			}
 		}
 		
 		break;
